chore: Adressing AI suggestions
